First Commit - Setup and Changes
This is a decent sized push. There was a bit of setup to get it working, but here we have it.  Over about 3.5-4 hours, I implemented the following changes:

1) Added page numbers
2) Each multiple choice answer on new line
3) Dates now fit on one row
4) Generate Header/Footer in own file
5) Header/Footer now are grey, rather than black.
6) Header/Footer take less space
7) Page margins now changed to better reflect Paper CRF
8) Opener fits on 1 page
9) Adjusted some font sizes

Overall, I think the changes look good and better reflect the style of the Paper CRFs.  However, much is to be done.

The largest, and most structural change is going to be adjusting the widths of columns.  At the moment, they are adjusted to accommodate full dates, as requested by our reviewers.  However, we will need to implement a much more nuanced system to facilitate questions that take up the whole width or vary in whether the question is wide or the answer.