Big Update
First round of comments responded to through a huge overhaul of the codebase.

The entire system has been refactored so that now, each form is split into Sections, Subsections, Rows, and Fields. These 4 are classes with methods to perform any necessary splitting and formatting directly in the class at the level that is appropriate.

Many small styles have been changed, questions grouped into subsections by 'branching logic', and any number of questions with any width column can be added.

Despite the flexibility of it, I found that structure looks better, so it is always 1, 2, or 3 questions per row with widths split into 6ths of the total form width.